# Allocate op arguments on stack when possible.