```
$ go test -run=$^ -bench=. -count=5 | tee /tmp/baseline.txt
$ go test -run=$^ -bench=. -count=5 | tee /tmp/experiment.txt
$ benchstat /tmp/{baseline,experiment}.txt
name                                 old time/op    new time/op    delta
NewWorkload/NoCallsNoGen-128            461ns ± 1%     458ns ± 1%    ~     (p=0.151 n=5+5)
NewWorkload/NoCalls-128                1.70µs ± 2%    1.68µs ± 1%    ~     (p=0.421 n=5+5)
NewWorkload/OneCall-128                1.75µs ± 1%    1.74µs ± 2%    ~     (p=0.310 n=5+5)
ResetExecutor/NoCallsNoGen-128         3.87µs ± 2%    3.88µs ± 2%    ~     (p=0.841 n=5+5)
ResetExecutor/NoCalls-128              5.26µs ± 2%    5.26µs ± 2%    ~     (p=0.643 n=5+5)
ResetExecutor/OneCall-128              5.60µs ± 3%    5.47µs ± 1%  -2.36%  (p=0.032 n=5+5)
Executions/NoCallsNoGen-128            12.4µs ± 2%    11.8µs ± 2%  -5.12%  (p=0.008 n=5+5)
Executions/NoCalls-128                 14.8µs ± 2%    14.0µs ± 2%  -5.45%  (p=0.008 n=5+5)
Executions/OneCall-128                 28.5µs ± 1%    27.5µs ± 1%  -3.72%  (p=0.008 n=5+5)
ParallelExecutions/NoCallsNoGen-128    1.78µs ± 5%    1.73µs ± 7%    ~     (p=0.222 n=5+5)
ParallelExecutions/NoCalls-128         1.89µs ± 1%    1.88µs ± 1%    ~     (p=0.389 n=5+5)
ParallelExecutions/OneCall-128         2.80µs ± 1%    2.81µs ± 2%    ~     (p=0.730 n=5+5)

name                                 old alloc/op   new alloc/op   delta
NewWorkload/NoCallsNoGen-128            0.00B          0.00B         ~     (all equal)
NewWorkload/NoCalls-128                  264B ± 0%      264B ± 0%    ~     (all equal)
NewWorkload/OneCall-128                  280B ± 0%      280B ± 0%    ~     (all equal)
ResetExecutor/NoCallsNoGen-128           576B ± 0%      576B ± 0%    ~     (all equal)
ResetExecutor/NoCalls-128                840B ± 0%      840B ± 0%    ~     (all equal)
ResetExecutor/OneCall-128                904B ± 0%      904B ± 0%    ~     (all equal)
Executions/NoCallsNoGen-128            1.10kB ± 0%    1.05kB ± 0%  -4.38%  (p=0.008 n=5+5)
Executions/NoCalls-128                 1.47kB ± 0%    1.39kB ± 0%  -5.47%  (p=0.008 n=5+5)
Executions/OneCall-128                 2.82kB ± 0%    2.74kB ± 0%  -2.85%  (p=0.000 n=5+4)
ParallelExecutions/NoCallsNoGen-128    1.10kB ± 0%    1.05kB ± 0%  -4.38%  (p=0.008 n=5+5)
ParallelExecutions/NoCalls-128         1.47kB ± 0%    1.39kB ± 0%  -5.43%  (p=0.008 n=5+5)
ParallelExecutions/OneCall-128         2.81kB ± 0%    2.73kB ± 0%  -2.84%  (p=0.000 n=5+4)

name                                 old allocs/op  new allocs/op  delta
NewWorkload/NoCallsNoGen-128             0.00           0.00         ~     (all equal)
NewWorkload/NoCalls-128                  6.00 ± 0%      6.00 ± 0%    ~     (all equal)
NewWorkload/OneCall-128                  7.00 ± 0%      7.00 ± 0%    ~     (all equal)
ResetExecutor/NoCallsNoGen-128           19.0 ± 0%      19.0 ± 0%    ~     (all equal)
ResetExecutor/NoCalls-128                25.0 ± 0%      25.0 ± 0%    ~     (all equal)
ResetExecutor/OneCall-128                27.0 ± 0%      27.0 ± 0%    ~     (all equal)
Executions/NoCallsNoGen-128              34.0 ± 0%      33.0 ± 0%  -2.94%  (p=0.008 n=5+5)
Executions/NoCalls-128                   44.0 ± 0%      43.0 ± 0%  -2.27%  (p=0.008 n=5+5)
Executions/OneCall-128                   78.0 ± 0%      77.0 ± 0%  -1.28%  (p=0.008 n=5+5)
ParallelExecutions/NoCallsNoGen-128      34.0 ± 0%      33.0 ± 0%  -2.94%  (p=0.008 n=5+5)
ParallelExecutions/NoCalls-128           44.0 ± 0%      43.0 ± 0%  -2.27%  (p=0.008 n=5+5)
ParallelExecutions/OneCall-128           78.0 ± 0%      77.0 ± 0%  -1.28%  (p=0.008 n=5+5)
```